[BugFix] build error fix
