+++ Test/Test.vcxproj
@@ -36,16 +36,13 @@
     <ClInclude Include="pch.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="add_test.cpp" />
-    <ClCompile Include="DelCommandTest.cpp" />
-    <ClCompile Include="intergration_test.cpp" />
-    <ClCompile Include="mod_test.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="test.cpp" />
     <ClCompile Include="test_sch1.cpp" />
     <ClCompile Include="test_sch2.cpp" />
   </ItemGroup>
